refactor: separate transform function types into a new file
