+++ packages/superset-ui-chart/src/components/SuperChart.tsx
@@ -5,7 +5,8 @@ import getChartTransformPropsRegistry from '../registries/ChartTransformPropsReg
 import ChartProps from '../models/ChartProps';
 import createLoadableRenderer, { LoadableRenderer } from './createLoadableRenderer';
 import { ChartType } from '../models/ChartPlugin';
-import { PreTransformProps, TransformProps, PostTransformProps } from '../types/Query';
+import { PreTransformProps, TransformProps, PostTransformProps } from '../types/TransformFunction';
+import { HandlerFunction } from '../types/Base';
 
 const IDENTITY = (x: any) => x;
 
@@ -22,8 +23,6 @@ const defaultProps = {
   onRenderFailure() {},
 };
 /* eslint-enable sort-keys */
-
-type HandlerFunction = (...args: any[]) => void;
 
 interface LoadingProps {
   error: any;
