+++ packages/superset-ui-chart/src/types/Base.ts
@@ -1,0 +1,5 @@
+export type HandlerFunction = (...args: any[]) => void;
+
+export interface PlainObject {
+  [key: string]: any;
+}
