+++ packages/superset-ui-chart/src/types/Query.ts
@@ -1,8 +1,6 @@
 /* eslint camelcase: 0 */
 import { DatasourceType } from './Datasource';
-import { ChartFormData } from './ChartFormData';
 import { AdhocMetric } from './Metric';
-import ChartProps from '../models/ChartProps';
 import { BinaryOperator, SetOperator, UnaryOperator } from './Operator';
 import { TimeRange } from './Time';
 
@@ -83,15 +81,3 @@ export interface QueryContext {
   };
   queries: QueryObject[];
 }
-
-export interface PlainProps {
-  [key: string]: any;
-}
-
-type TransformFunction<Input = PlainProps, Output = PlainProps> = (x: Input) => Output;
-
-export type PreTransformProps = TransformFunction<ChartProps, ChartProps>;
-export type TransformProps = TransformFunction<ChartProps>;
-export type PostTransformProps = TransformFunction;
-
-export type BuildQueryFunction<T extends ChartFormData> = (formData: T) => QueryContext;
